+++ data/src/main/java/com/linku/data/api/LinkuApi.kt
@@ -7,7 +7,7 @@ import com.linku.data.api.dto.server.LinkuResultDTO
 import com.linku.data.api.dto.server.LinkuSimpleDTO
 import com.linku.data.api.dto.server.LinkuUpdateDTO
 import com.linku.data.api.dto.server.QuickSearchResult
-import com.linku.data.api.dto.server.UpdateLinkFolderDTO
+import com.linku.data.api.dto.folder.UpdateLinkFolderDTO
 import okhttp3.MultipartBody
 import okhttp3.RequestBody
 import retrofit2.Response
