+++ data/src/main/java/com/linku/data/api/dto/folder/FolderCreateRequestDTO.kt
@@ -1,11 +1,11 @@
-package com.linku.data.api.dto.server
+package com.linku.data.api.dto.folder
 
 import com.squareup.moshi.Json
 
 import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
-data class FolderUpdateRequestDTO(
+data class FolderCreateRequestDTO(
 
     @field:Json(name = "folderName")
     val folderName: String
