+++ data/src/main/java/com/linku/data/api/dto/folder/FolderListResponseDTO.kt
@@ -1,11 +1,12 @@
-package com.linku.data.api.dto.server
+package com.linku.data.api.dto.folder
 
 import com.squareup.moshi.Json
 
+// com.linku.core.model.FolderSimpleInfo -> 구조를 공유
 import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
-data class FolderTreeResponseDTO(
+data class FolderListResponseDTO(
 
     @field:Json(name = "folderId")
     val folderId: Long,
@@ -13,12 +14,12 @@ data class FolderTreeResponseDTO(
     @field:Json(name = "folderName")
     val folderName: String,
 
-    @field:Json(name = "categoryId")
-    val categoryId: Long,
+    @field:Json(name = "parentFolderId")
+    val parentFolderId: Long?,
 
     @field:Json(name = "isBookmarked")
     val isBookmarked: Boolean,
 
-    @field:Json(name = "children")
-    val children: List<FolderTreeResponseDTO>? = null
+    @field:Json(name = "isSharing")
+    val isSharing: String? = ""
 )
