+++ data/src/main/java/com/linku/data/api/dto/folder/FolderTreeResponseDTO.kt
@@ -1,11 +1,11 @@
-package com.linku.data.api.dto.server
+package com.linku.data.api.dto.folder
 
 import com.squareup.moshi.Json
 
 import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
-data class GetParentFoldersDTO(
+data class FolderTreeResponseDTO(
 
     @field:Json(name = "folderId")
     val folderId: Long,
@@ -13,9 +13,12 @@ data class GetParentFoldersDTO(
     @field:Json(name = "folderName")
     val folderName: String,
 
-    @field:Json(name = "parentFolderId")
-    val parentFolderId: Long?,
+    @field:Json(name = "categoryId")
+    val categoryId: Long,
 
     @field:Json(name = "isBookmarked")
     val isBookmarked: Boolean,
+
+    @field:Json(name = "children")
+    val children: List<FolderTreeResponseDTO>? = null
 )
