+++ data/src/main/java/com/linku/data/api/dto/folder/FolderUpdateRequestDTO.kt
@@ -1,11 +1,11 @@
-package com.linku.data.api.dto.server
+package com.linku.data.api.dto.folder
 
 import com.squareup.moshi.Json
 
 import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
-data class FolderCreateRequestDTO(
+data class FolderUpdateRequestDTO(
 
     @field:Json(name = "folderName")
     val folderName: String
