+++ data/src/main/java/com/linku/data/implementation/repository/FolderRepositoryImpl.kt
@@ -10,12 +10,12 @@ import com.linku.core.model.SharedFolderInfo
 import com.linku.core.model.SharedFolderSimpleInfo
 import com.linku.core.repository.FolderRepository
 import com.linku.data.api.ServerApi
-import com.linku.data.api.dto.server.FolderCreateRequestDTO
-import com.linku.data.api.dto.server.FolderTreeResponseDTO
-import com.linku.data.api.dto.server.FolderUpdateRequestDTO
-import com.linku.data.api.dto.server.LinksFoldersResponseDTO
-import com.linku.data.api.dto.server.UpdateBookmarkRequestDTO
-import com.linku.data.api.dto.server.UpdateLinkFolderDTO
+import com.linku.data.api.dto.folder.FolderCreateRequestDTO
+import com.linku.data.api.dto.folder.FolderTreeResponseDTO
+import com.linku.data.api.dto.folder.FolderUpdateRequestDTO
+import com.linku.data.api.dto.folder.LinksFoldersResponseDTO
+import com.linku.data.api.dto.folder.UpdateBookmarkRequestDTO
+import com.linku.data.api.dto.folder.UpdateLinkFolderDTO
 import com.linku.data.api.withAuth
 import com.linku.data.api.withAuthResp204Raw
 import com.linku.data.preference.AuthPreference
